[refactor] moved prompt dialog fragment directory
