+++ frontend/app/src/main/java/com/project/spire/ui/create/PromptDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.project.spire.ui.auth
+package com.project.spire.ui.create
 
 import android.app.Dialog
 import android.os.Bundle
